axios

+++ instruction/02-fetch/js/script.js
@@ -1,7 +1,60 @@
-import { getAllUsers, getUserById, deleteUserById } from "./users-api.js";
+import {
+  getAllUsers,
+  getUserById,
+  deleteUserById,
+  createUser,
+  updateUser,
+} from "./users-api.js";
 
 const lstUsers = document.getElementById("lstUsers");
 const userDetails = document.getElementById("userDetails");
+const frmUser = document.getElementById("frmUser");
+const txtFirstName = document.getElementById("firstName");
+const txtLastName = document.getElementById("lastName");
+const txtEmail = document.getElementById("email");
+const btnSubmit=document.getElementById("btnSubmit");
+
+const addUser = async () => {
+  const firstName = txtFirstName.value;
+  const lastName = txtLastName.value;
+  const email = txtEmail.value;
+
+  const user = {
+    firstName,
+    lastName,
+    email,
+  };
+  await createUser(user);
+  init();
+  frmUser.reset();
+};
+
+const editUser = async (id) => {
+  const user=await getUserById(id);
+  const{firstName,lastName,email}=user;
+  txtFirstName.value=firstName;
+  txtLastName.value=lastName;
+  txtEmail.value=email;
+
+};
+
+const saveUser=async (id) => {
+  const firstName = txtFirstName.value;
+  const lastName = txtLastName.value;
+  const email = txtEmail.value;
+
+  const user = {
+    firstName,
+    lastName,
+    email,
+  };
+  await updateUser(id,user);
+  init();
+  frmUser.reset();
+  frmUser.dataset.method="create";
+  btnSubmit.innerHTML="➕ Add";
+
+};
 
 const hiddenPassword = (password) => {
   return "*".repeat(password.length);
@@ -21,7 +74,8 @@ const renderUsers = (users) => {
                           <li class="list-group-item">Last Name: ${item.lastName}</li>
                         </ul> 						            
                         <button class="btn btn-danger btn-del" style="cursor: pointer;" data-id="${item.id}"> 🗑️ Delete</button>
-                      </div>
+                      <button class="btn btn-info my-3 btn-edit" type="submit" data-id="${item.id}">✍️ Edit</button>
+                        </div>
                     </div>
                   </div>`;
   });
@@ -30,7 +84,8 @@ const renderUsers = (users) => {
 };
 
 const renderUserDetails = (user) => {
-  const { firstName, lastName, email, phone, userName, password, avatar } = user;
+  const { firstName, lastName, email, phone, userName, password, avatar } =
+    user;
   const hiddenPwd = hiddenPassword(password);
 
   const strUser = `<div class="col-md-5">
@@ -65,7 +120,7 @@ init();
 lstUsers.addEventListener("click", async (e) => {
   if (e.target.closest(".card") && !e.target.classList.contains("btn-del")) {
     const card = e.target.closest(".card");
-    const userId = card.querySelector('.btn-del').dataset.id;
+    const userId = card.querySelector(".btn-del").dataset.id;
     console.log(`Fetching details for user ID: ${userId}`);
 
     try {
@@ -80,12 +135,12 @@ lstUsers.addEventListener("click", async (e) => {
 });
 
 lstUsers.addEventListener("click", async (e) => {
-  if (e.target.classList.contains("btn-del")) {
-    const res=confirm("Are you sure to delete");
-    if(!res) return;
-    const userId = e.target.dataset.id;
-    console.log(`Deleting user with ID: ${userId}`);
+  const userId = e.target.dataset.id;
+  if (!userId) return;
 
+  if (e.target.classList.contains("btn-del")) {
+    const res = confirm("Are you sure to delete");
+    if (!res) return;
     try {
       await deleteUserById(userId);
       console.log(`User with ID: ${userId} deleted successfully`);
@@ -93,5 +148,27 @@ lstUsers.addEventListener("click", async (e) => {
     } catch (error) {
       console.error("Failed to delete user", error);
     }
+  } else if (e.target.classList.contains("btn-edit")) {
+  
+    editUser(userId);
+    window.scrollTo(0,0);
+    btnSubmit.innerHTML="🖫 Update";
+    frmUser.dataset.method="update";
+    frmUser.dataset.id=userId;
+    init();
+
   }
+});
+
+frmUser.addEventListener("submit", async (e) => {
+  e.preventDefault();
+  const method=e.target.dataset.method;
+  const id=e.target.dataset.id;
+
+  if(method ==="create"){
+    addUser();
+  }else{
+    saveUser(id);
+  }
+  
 });
